refactor: api route와 헷갈리지 않도록 app 외부 api 디렉터리 이름을 apis로 변경
