+++ client/src/app/api/post-document/route.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import {PostDocumentContent} from '@api/document';
+import {PostDocumentContent} from '@apis/document';
 import {CACHE} from '@constants/cache';
 import {ENDPOINT} from '@constants/endpoint';
 import {WikiDocument} from '@type/Document.type';
